+++ tests/test_fixef.R
@@ -11,9 +11,11 @@
 ############# (1): general run tests #############
 library(plm)
 data("Grunfeld", package = "plm")
-gi <- plm(inv ~ value + capital, data = Grunfeld, model = "within")
-gt <- plm(inv ~ value + capital, data = Grunfeld, model = "within", effect = "time")
-gd <- plm(inv ~ value + capital, data = Grunfeld, model = "within", effect = "twoways")
+
+# balanced models
+gi  <- plm(inv ~ value + capital, data = Grunfeld, model = "within")
+gt  <- plm(inv ~ value + capital, data = Grunfeld, model = "within", effect = "time")
+gtw <- plm(inv ~ value + capital, data = Grunfeld, model = "within", effect = "twoways")
 
 
 f_level             <- fixef(gi, type = "level")
@@ -35,8 +37,8 @@ print(attr(f_level_t_robust_func, "se"))
 print(summary(f_level_t),             digits = 8)
 print(summary(f_level_t_robust_func), digits = 8)
 
-f_level_d             <- fixef(gd, type = "level")
-f_level_d_robust_func <- fixef(gd, type = "level", vcov = vcovHC)     # vcov is function
+f_level_d             <- fixef(gtw, type = "level")
+f_level_d_robust_func <- fixef(gtw, type = "level", vcov = vcovHC)     # vcov is function
 
 print(attr(f_level_d,             "se"))
 print(attr(f_level_d_robust_func, "se"))
@@ -47,14 +49,18 @@ print(summary(f_level_d_robust_func), digits = 8)
 # just run tests for type = "dmean" and type = "dfirst"
 fixef(gi, type = "dmean")
 fixef(gt, type = "dmean")
-fixef(gd, effect = "individual", type = "dmean")
-fixef(gd, effect = "time",       type = "dmean")
+fixef(gtw, effect = "individual", type = "dmean")
+fixef(gtw, effect = "time",       type = "dmean")
+fixef(gtw, effect = "twoways",    type = "dmean")
+
 
 fixef(gi, type = "dfirst")
 fixef(gt, type = "dfirst")
-fixef(gd, effect = "individual", type = "dfirst")
-fixef(gd, effect = "time",       type = "dfirst")
+fixef(gtw, effect = "individual", type = "dfirst")
+fixef(gtw, effect = "time",       type = "dfirst")
+fixef(gtw, effect = "twoways",    type = "dfirst")
 
+fixef(gtw, effect = "twoways",    type = "level")
 
 ############# (2): consistency with summary.plm #############
 # compare summary.plm to summary.fixef( , type = "dfirst")
@@ -63,23 +69,23 @@ sum_mod_pool <- summary(mod_pool)
 f_dfirst <- fixef(gi, type = "dfirst")
 sum_f_dfirst <- summary(f_dfirst)
 
-if (!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Estimate"], sum_f_dfirst[ , "Estimate"], check.attributes = FALSE)))
+if(!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Estimate"], sum_f_dfirst[ , "Estimate"], check.attributes = FALSE)))
   stop("estimates diverge: summary.plm vs. summary.fixef(..., type = \"dfirst\")")
 
-if (!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Std. Error"], sum_f_dfirst[ , "Std. Error"], check.attributes = FALSE)))
+if(!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Std. Error"], sum_f_dfirst[ , "Std. Error"], check.attributes = FALSE)))
   stop("standard errors diverge: summary.plm vs. summary.fixef(..., type = \"dfirst\")")
 
-if (!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "t-value"], sum_f_dfirst[ , "t-value"], check.attributes = FALSE)))
+if(!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "t-value"], sum_f_dfirst[ , "t-value"], check.attributes = FALSE)))
   stop("t-values diverge: summary.plm vs. summary.fixef(..., type = \"dfirst\")")
 
-if (!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Pr(>|t|)"], sum_f_dfirst[ , "Pr(>|t|)"], check.attributes = FALSE)))
+if(!isTRUE(all.equal(sum_mod_pool[["coefficients"]][-c(1:3) , "Pr(>|t|)"], sum_f_dfirst[ , "Pr(>|t|)"], check.attributes = FALSE)))
   stop("p-values diverge: summary.plm vs. summary.fixef(..., type = \"dfirst\")")
 
 
 ###### compare to package lfe:
 ## Standard errors are bootstrapped in lfe
-##  -> different results compared to plm
-##  -> different results for every new call
+##  -> different SE results compared to plm
+##  -> different SE results for every new call
 #
 # library(lfe)
 # data("Grunfeld", package = "plm")